+++ pkg/gatewayserver/io/udp/firewall_test.go
@@ -21,12 +21,15 @@ import (
 	"time"
 
 	"github.com/smartystreets/assertions"
+	"go.thethings.network/lorawan-stack/pkg/errors"
 	. "go.thethings.network/lorawan-stack/pkg/gatewayserver/io/udp"
 	encoding "go.thethings.network/lorawan-stack/pkg/ttnpb/udp"
 	"go.thethings.network/lorawan-stack/pkg/types"
 	"go.thethings.network/lorawan-stack/pkg/util/test"
 	"go.thethings.network/lorawan-stack/pkg/util/test/assertions/should"
 )
+
+func isNoError(err error) bool { return err == nil }
 
 func TestMemoryFirewall(t *testing.T) {
 	ctx := test.Context()
@@ -58,23 +61,23 @@ func TestMemoryFirewall(t *testing.T) {
 	}
 
 	for i, tc := range []struct {
-		Packet    encoding.Packet
-		OK        bool
-		WaitAfter time.Duration
+		Packet     encoding.Packet
+		ErrorCheck func(error) bool
+		WaitAfter  time.Duration
 	}{
 		{
 			Packet: encoding.Packet{
 				GatewayAddr: &downAddr1,
 				PacketType:  encoding.PullData,
 			},
-			OK: false, // no EUI
+			ErrorCheck: errors.IsInvalidArgument, // no EUI
 		},
 		{
 			Packet: encoding.Packet{
 				GatewayEUI: &eui1,
 				PacketType: encoding.PullData,
 			},
-			OK: false, // no address
+			ErrorCheck: errors.IsInvalidArgument, // no address
 		},
 		{
 			Packet: encoding.Packet{
@@ -82,7 +85,7 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &downAddr1,
 				PacketType:  encoding.PullData,
 			},
-			OK: true, // downstream 1
+			ErrorCheck: isNoError, // downstream 1
 		},
 		{
 			Packet: encoding.Packet{
@@ -90,7 +93,7 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &downAddr1,
 				PacketType:  encoding.PullData,
 			},
-			OK: true, // second time downstream 1 with same address
+			ErrorCheck: isNoError, // second time downstream 1 with same address
 		},
 		{
 			Packet: encoding.Packet{
@@ -98,7 +101,7 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &upAddr1,
 				PacketType:  encoding.PushData,
 			},
-			OK: true, // upstream 1 with same address
+			ErrorCheck: isNoError, // upstream 1 with same address
 		},
 		{
 			Packet: encoding.Packet{
@@ -106,7 +109,7 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &downAddr2,
 				PacketType:  encoding.PullData,
 			},
-			OK: true, // first time downlink from 2
+			ErrorCheck: isNoError, // first time downlink from 2
 		},
 		{
 			Packet: encoding.Packet{
@@ -114,7 +117,7 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &upAddr2,
 				PacketType:  encoding.PushData,
 			},
-			OK: true, // first time uplink from 2
+			ErrorCheck: isNoError, // first time uplink from 2
 		},
 		{
 			Packet: encoding.Packet{
@@ -122,8 +125,8 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &addr3,
 				PacketType:  encoding.PullData,
 			},
-			WaitAfter: block * 2,
-			OK:        false, // block change of downlink address
+			WaitAfter:  block * 2,
+			ErrorCheck: errors.IsFailedPrecondition, // block change of downlink address
 		},
 		{
 			Packet: encoding.Packet{
@@ -131,14 +134,14 @@ func TestMemoryFirewall(t *testing.T) {
 				GatewayAddr: &addr3,
 				PacketType:  encoding.PullData,
 			},
-			OK: true, // permit change of downlink address after block time
+			ErrorCheck: isNoError, // permit change of downlink address after block time
 		},
 	} {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			a := assertions.New(t)
 
-			actual := v.Filter(tc.Packet)
-			if !a.So(actual, should.Equal, tc.OK) {
+			err := v.Filter(tc.Packet)
+			if !a.So(tc.ErrorCheck(err), should.BeTrue) {
 				t.FailNow()
 			}