+++ pkg/version/ttn.go
@@ -3,4 +3,4 @@
 package version
 
 // TTN Version
-var TTN = "v3.3.0-dev"
+var TTN = "v3.3.1-dev"